refactor: extract pane tree model
