+++ scripts/rust-source-size-baseline.txt
@@ -2,5 +2,4 @@
 # A file may only stay equal to this maximum or shrink with this entry tightened.
 crates/shell-core/src/actions.rs 1591
 crates/shell-core/src/control.rs 1183
-crates/shell-core/src/model.rs 1320
 crates/shell-core/src/reducer.rs 2723
